Revert "[Reland][X86] Re-enable some inductor-related test cases (#3777)"
This reverts commit 6ad7c40461a5c8e79e442f17d8c25b47f0ae652f.
